Complete flow for album deletion

+++ music-collector/src/app/fragments/album-list/album-list.component.ts
@@ -1,7 +1,10 @@
-import { Component, OnDestroy, OnInit } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { MusicService } from 'src/app/music.service';
+import { MatDialog } from '@angular/material/dialog';
+
 
 import { Album } from '../../model/album.model';
+import { DeleteDialogComponent } from '../delete-dialog/delete-dialog.component';
 
 @Component({
   selector: 'app-album-list',
@@ -13,7 +16,9 @@ export class AlbumListComponent implements OnInit {
   albumList: Album[] = [];
   loaded: boolean = false;
 
-  constructor(private musicService: MusicService) { }
+  constructor(
+    private musicService: MusicService,
+    private dialog: MatDialog) { }
 
   ngOnInit(): void {
     this.musicService.albumListStream().subscribe((newAlbumList: Album[]) => {
@@ -30,10 +35,24 @@ export class AlbumListComponent implements OnInit {
     this.musicService.retrieveAlbumsFromDB();
   }
 
-  deleteAlbum(album): void {
-    this.musicService.deleteAlbum(album.id).subscribe(() => {
-      this.getAlbumList();
+  openDeleteAlbumDialog(album: Album): void {
+    let dialogRef = this.dialog.open(
+      DeleteDialogComponent,
+      {
+        data: {
+          album: album
+        }
+      });
+
+    dialogRef.afterClosed().subscribe((deletionConfirmed) => {
+      if (deletionConfirmed) {
+        this.musicService.deleteAlbum(album.id).subscribe(() => {
+          this.getAlbumList();
+        })
+      }
     })
+
+
   }
